players/smart-hearts/requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
PySimpleGUI==4.60.4
python-dateutil==2.8.2
six==1.16.0
zipp==3.15.0
